# src/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Doormat",
    "description": "AI-first rental finder - autonomous property manager discovery and listing scoring",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "summary": "List Preferences",
        "description": "Return saved search preferences, newest first.",
        "operationId": "list_preferences_api_preferences_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreferenceResponse"
                  },
                  "title": "Response List Preferences Api Preferences Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Create Preference",
        "description": "Create a natural-language search preference.",
        "operationId": "create_preference_api_preferences_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences/{preference_id}": {
      "patch": {
        "tags": [
          "preferences"
        ],
        "summary": "Update Preference",
        "description": "Update an existing search preference.",
        "operationId": "update_preference_api_preferences__preference_id__patch",
        "parameters": [
          {
            "name": "preference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preference Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "preferences"
        ],
        "summary": "Delete Preference",
        "description": "Delete a saved search preference.",
        "operationId": "delete_preference_api_preferences__preference_id__delete",
        "parameters": [
          {
            "name": "preference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preference Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreferenceCreate": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 10,
            "title": "Description"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "City"
          },
          "api_provider": {
            "type": "string",
            "maxLength": 50,
            "title": "Api Provider",
            "default": "openrouter"
          },
          "openrouter_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Api Key"
          },
          "apify_api_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Apify Api Token"
          },
          "fast_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Fast Model"
          },
          "smart_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Smart Model"
          }
        },
        "type": "object",
        "required": [
          "description",
          "city"
        ],
        "title": "PreferenceCreate",
        "description": "Create a new preference."
      },
      "PreferenceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "api_provider": {
            "type": "string",
            "title": "Api Provider"
          },
          "has_openrouter_api_key": {
            "type": "boolean",
            "title": "Has Openrouter Api Key",
            "default": false
          },
          "openrouter_key_last4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Key Last4"
          },
          "has_apify_api_token": {
            "type": "boolean",
            "title": "Has Apify Api Token",
            "default": false
          },
          "apify_token_last4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apify Token Last4"
          },
          "fast_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fast Model"
          },
          "smart_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smart Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "city",
          "api_provider",
          "created_at",
          "updated_at"
        ],
        "title": "PreferenceResponse",
        "description": "Preference response model."
      },
      "PreferenceUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "api_provider": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Provider"
          },
          "openrouter_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Api Key"
          },
          "apify_api_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Apify Api Token"
          },
          "fast_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Fast Model"
          },
          "smart_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Smart Model"
          }
        },
        "type": "object",
        "title": "PreferenceUpdate",
        "description": "Update an existing preference."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}